refactor: isolate batch GraphQL compilation
